💯test: 게시글 수정 테스트 API 수정
